+++ packages/component/src/provider/appProvider.tsx
@@ -24,6 +24,7 @@ import {
 import toast, { useToasterStore } from "react-hot-toast";
 import { LocalProvider } from "@/i18n";
 import { IContract } from "@orderly.network/core";
+import { praseChainIdToNumber } from "@orderly.network/utils";
 
 export type AppStateErrors = {
   ChainNetworkNotSupport: boolean;
@@ -156,29 +157,30 @@ const InnerProvider = (props: PropsWithChildren<OrderlyAppProviderProps>) => {
   });
 
   const checkChainId = useCallback(
-    (chainId: string): boolean => {
+    (chainId: number): boolean => {
       if (!chainId || !chains) {
         return false;
       }
 
-      if (typeof chainId === "number") {
-        chainId = `0x${Number(chainId).toString(16)}`;
-      }
+      // if (typeof chainId === "number") {
+      //   chainId = `0x${Number(chainId).toString(16)}`;
+      // }
 
       //
 
       // check whether chain id and network id match
-      const chainIdNum = parseInt(chainId, 16);
+      // const chainIdNum = parseInt(chainId, 16);
       if (
-        (networkId === "mainnet" && chainIdNum === 421613) ||
-        (networkId === "testnet" && chainIdNum !== 421613 && chainIdNum !== 420)
+        (networkId === "mainnet" && chainId === 421613) ||
+        (networkId === "testnet" && chainId !== 421613)
       ) {
         return false;
       }
 
       const isSupport = chains.some((item: { id: string | number }) => {
-        if (typeof item.id === "number") {
-          return `0x${Number(item.id).toString(16)}` === chainId;
+        if (typeof item.id === "string") {
+          // return `0x${Number(item.id).toString(16)}` === chainId;
+          return parseInt(item.id, 16) === chainId;
         }
         return item.id === chainId;
       });
@@ -214,7 +216,9 @@ const InnerProvider = (props: PropsWithChildren<OrderlyAppProviderProps>) => {
         // account.address = wallet.accounts[0].address;
         const status = await account.setAddress(wallet.accounts[0].address, {
           provider: wallet.provider,
-          chain: wallet.chains[0],
+          chain: {
+            id: praseChainIdToNumber(wallet.chains[0].id),
+          },
           wallet: {
             name: wallet.label,
           },
@@ -230,7 +234,7 @@ const InnerProvider = (props: PropsWithChildren<OrderlyAppProviderProps>) => {
 
   const _onWalletDisconnect = useCallback(async (): Promise<any> => {
     if (typeof disconnect === "function" && currentWallet) {
-      console.warn("🤜 disconnect wallet");
+      console.log("🤜 disconnect wallet");
 
       return disconnect(currentWallet).then(() => {
         return account.disconnect();
@@ -256,19 +260,20 @@ const InnerProvider = (props: PropsWithChildren<OrderlyAppProviderProps>) => {
     return currentWallet.accounts[0].address;
   }, [currentWallet]);
 
-  const currentChainId = useMemo(() => {
+  // current connected chain id
+  const currentChainId = useMemo<number | null>(() => {
     if (!currentWallet) {
       return null;
     }
-    return currentWallet.chains[0].id;
+    const id = currentWallet.chains[0].id;
+
+    return praseChainIdToNumber(id);
   }, [currentWallet]);
 
   useEffect(() => {
     // currentWallet?.provider.detectNetwork().then((x) =>
 
-    console.log("chains", chains);
-
-    if (!chains || chains.length === 0) {
+    if (!chains || chains.length === 0 || !currentChainId) {
       return;
     }
 
@@ -287,10 +292,10 @@ const InnerProvider = (props: PropsWithChildren<OrderlyAppProviderProps>) => {
         // console.log("currentWallet 22 ", currentAddress, currentChainId);
         return;
       }
-      // 需要确定已经拿到chains list
+
       if (!checkChainId(currentChainId)) {
         // console.warn("!!!! not support this chian -> disconnect wallet");
-        // TODO: 确定是否需要断开连接
+
         // account.disconnect();
         // @ts-ignore
         setErrors((errors) => ({ ...errors, ChainNetworkNotSupport: true }));
@@ -306,7 +311,10 @@ const InnerProvider = (props: PropsWithChildren<OrderlyAppProviderProps>) => {
 
       account.setAddress(currentWallet.accounts[0].address, {
         provider: currentWallet.provider,
-        chain: currentWallet.chains[0],
+        chain: {
+          id: currentChainId,
+          // name: currentWallet.chains[0].name,
+        },
         wallet: {
           name: currentWallet.label,
         },