+++ packages/utils/src/chain.ts
@@ -4,3 +4,13 @@ export const praseChainId = (chainId: string | number): number => {
   if (typeof chainId === "string") return hex2int(chainId);
   return chainId;
 };
+
+export const praseChainIdToNumber = (chainId: string | number): number => {
+  if (
+    typeof chainId === "string" &&
+    chainId.startsWith("0x") &&
+    /^[a-f0-9]+$/iu.test(chainId.slice(2))
+  )
+    return hex2int(chainId);
+  return chainId as number;
+};