+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,57 @@
 # examples
 
+## 0.1.91
+
+### Patch Changes
+
+- 268abc5: fix bug
+- fix bugs
+- fix bugs
+- fix bug
+- Updated dependencies [268abc5]
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/react@0.1.30
+  - @orderly.network/web3-onboard@0.0.27
+  - @orderly.network/hooks@1.0.27
+
+## 0.1.91-alpha.1
+
+### Patch Changes
+
+- fix bug
+- Updated dependencies
+  - @orderly.network/react@0.1.30-alpha.1
+  - @orderly.network/web3-onboard@0.0.27-alpha.1
+  - @orderly.network/hooks@1.0.27-alpha.1
+
+## 0.1.91-alpha.0
+
+### Patch Changes
+
+- fix bugs
+- Updated dependencies
+  - @orderly.network/react@0.1.30-alpha.0
+  - @orderly.network/web3-onboard@0.0.27-alpha.0
+  - @orderly.network/hooks@1.0.27-alpha.0
+
+## 0.1.90
+
+### Patch Changes
+
+- 268abc5: fix bug
+- fix style issues
+- fix bug
+- fix bugs
+- Updated dependencies [268abc5]
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - @orderly.network/react@0.1.29
+  - @orderly.network/web3-onboard@0.0.26
+  - @orderly.network/hooks@1.0.26
+
 ## 0.1.90-alpha.2
 
 ### Patch Changes