+++ packages/types/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @orderly.network/types
 
+## 0.1.26
+
+### Patch Changes
+
+- 268abc5: fix bug
+- fix bugs
+- fix bugs
+- fix bug
+
+## 0.1.26-alpha.1
+
+### Patch Changes
+
+- fix bug
+
+## 0.1.26-alpha.0
+
+### Patch Changes
+
+- fix bugs
+
+## 0.1.25
+
+### Patch Changes
+
+- 268abc5: fix bug
+- fix style issues
+- fix bug
+- fix bugs
+
 ## 0.1.25-alpha.2
 
 ### Patch Changes